+++ packages/snaps-utils/src/json-rpc.ts
@@ -1,3 +1,5 @@
+import { KeyringCapabilitiesStruct as CapabilitiesStruct } from '@metamask/keyring-api';
+import type { KeyringCapabilities as Capabilities } from '@metamask/keyring-api';
 import { SubjectType } from '@metamask/permission-controller';
 import type { Infer } from '@metamask/superstruct';
 import {
@@ -95,6 +97,32 @@ export function assertIsKeyringOrigins(
     value,
     KeyringOriginsStruct,
     'Invalid keyring origins',
+    ErrorWrapper,
+  );
+}
+
+export const KeyringCapabilitiesStruct = object({
+  capabilities: optional(CapabilitiesStruct),
+});
+
+export type KeyringCapabilities = Infer<typeof KeyringCapabilitiesStruct>;
+
+/**
+ * Assert that the given value is a valid {@link KeyringCapabilities} object.
+ *
+ * @param value - The value to assert.
+ * @param ErrorWrapper - An optional error wrapper to use. Defaults to
+ * {@link AssertionError}.
+ * @throws If the value is not a valid {@link KeyringCapabilities} object.
+ */
+export function assertIsKeyringCapabilities(
+  value: unknown,
+  ErrorWrapper?: AssertionErrorConstructor,
+): asserts value is KeyringCapabilities {
+  assertStruct(
+    value,
+    KeyringCapabilitiesStruct,
+    'Invalid keyring capabilities',
     ErrorWrapper,
   );
 }