+++ packages/snaps-utils/src/manifest/validation.ts
@@ -33,7 +33,11 @@ import {
 import { isDerivationPathEqual } from '../array';
 import { CronjobSpecificationArrayStruct } from '../cronjob';
 import { SIP_6_MAGIC_VALUE, STATE_ENCRYPTION_MAGIC_VALUE } from '../entropy';
-import { KeyringOriginsStruct, RpcOriginsStruct } from '../json-rpc';
+import {
+  KeyringCapabilitiesStruct,
+  KeyringOriginsStruct,
+  RpcOriginsStruct,
+} from '../json-rpc';
 import { SnapIdStruct } from '../snaps';
 import { mergeStructs, type InferMatching } from '../structs';
 import { NameStruct, NpmSnapFileNames, uri } from '../types';
@@ -216,7 +220,11 @@ export const PermissionsStruct: Describe<InitialPermissions> = type({
   ),
   'endowment:ethereum-provider': optional(EmptyObjectStruct),
   'endowment:keyring': optional(
-    mergeStructs(HandlerCaveatsStruct, KeyringOriginsStruct),
+    mergeStructs(
+      HandlerCaveatsStruct,
+      KeyringOriginsStruct,
+      KeyringCapabilitiesStruct,
+    ),
   ),
   'endowment:protocol': optional(
     mergeStructs(