+++ packages/snaps-utils/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add capabilities caveat for `endowment:keyring` ([#XXXX](https://github.com/MetaMask/snaps/pull/XXXX))
+  - `KeyringCapabilities` and `KeyringCapabilitiesStruct` from `@metamask/keyring-api` are now re-exported from this package as well.
+
 ## [12.1.1]
 
 ### Changed